signup connected to db
